Make products service to divide view and logic
